pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "helical"
version = "0.0.1a28"
authors = [
  { name="Helical Team", email="support@helical-ai.com" },
]
description = "Helical Python SDK"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
]

license = {file = "LICENSE"}

dependencies = [
    'requests>=2.32.3',
    'pandas==2.2.2',
    'anndata==0.11',
    'numpy==1.26.4',
    'scikit-learn>=1.5.0',
    'scipy==1.13.1',
    'gitpython==3.1.43',
    'torch==2.6.0',
    'torchvision==0.21.0',
    'accelerate==1.4.0',
    'transformers==4.49.0',
    'loompy==3.0.7',
    'scib==1.1.5',
    'scikit-misc==0.3.1',
    'azure-identity==1.16.1',
    'azure-storage-blob==12.19.1',
    'azure-core==1.30.1',
    'einops==0.8.0',
    'omegaconf==2.3.0',
    'hydra-core==1.3.2',
    'louvain==0.8.2',
    'pyensembl',
    'datasets==2.20.0',
]


[project.optional-dependencies]
mamba-ssm = [
    'mamba-ssm==2.2.4',
    'causal-conv1d==1.5.0.post8',
]

evo-2 = [
    'biopython==1.85',
    'transformer-engine[pytorch]==1.13.0',
]

[tool.hatch.metadata]
allow-direct-references = true # Allows installing package from git

[project.urls]
Homepage = "https://github.com/helicalAI/helical"
Issues = "https://github.com/helicalAI/helical/issues"
Documentation = "https://helical.readthedocs.io/"
Repository = "https://github.com/helicalAI/helical"
